Pull up default routes for /openapi and /version
