+++ src/azul/openapi/spec.py
@@ -173,62 +173,3 @@ class CommonEndpointSpecs:
                 ]
             }
         }
-
-    @property
-    def openapi(self) -> JSON:
-        return {
-            'method_spec': {
-                'summary': 'Return OpenAPI specifications for this REST API',
-                'description': format_description('''
-                    This endpoint returns the [OpenAPI specifications]'
-                    (https://github.com/OAI/OpenAPI-Specification) for this REST
-                    API. These are the specifications used to generate the page
-                    you are visiting now.
-                '''),
-                'responses': {
-                    '200': {
-                        'description': '200 response',
-                        **responses.json_content(
-                            schema.object(
-                                openapi=str,
-                                **{
-                                    k: schema.object()
-                                    for k in ('info', 'tags', 'servers', 'paths', 'components')
-                                }
-                            )
-                        )
-                    }
-                },
-                'tags': ['Auxiliary']
-            }
-        }
-
-    @property
-    def version(self) -> JSON:
-        return {
-            'method_spec': {
-                'summary': 'Describe current version of this REST API',
-                'tags': ['Auxiliary'],
-                'responses': {
-                    '200': {
-                        'description': 'Version endpoint is reachable.',
-                        **responses.json_content(
-                            schema.object(
-                                git=schema.object(
-                                    commit=str,
-                                    dirty=bool
-                                ),
-                                changes=schema.array(
-                                    schema.object(
-                                        title=str,
-                                        issues=schema.array(str),
-                                        upgrade=schema.array(str),
-                                        notes=schema.optional(str)
-                                    )
-                                )
-                            )
-                        )
-                    }
-                }
-            }
-        }
